add product ui completed

+++ frontend-admin-app/src/actions/auth.actions.js
@@ -1,23 +1,27 @@
-import { authConstants } from "./constants";
+import { authConstants, userConstants } from "./constants";
 import axios from '../helpers/axios';
 
 export const login = (user) => {
     return async (dispatch) => {
-        dispatch({type: authConstants.LOGIN_REQUEST});
-        const res = await axios.post('/admin/signin', {...user});
+        dispatch({ type: authConstants.LOGIN_REQUEST });
+        const res = await axios.post('/admin/signin', { ...user });
 
-        if(res.status === 200) {
-            const { token, user} = res.data;
+        if (res.status === 200) {
+            const { token, user } = res.data;
             localStorage.setItem('token', token);
             localStorage.setItem('user', JSON.stringify(user));
-            dispatch({type: authConstants.LOGIN_SUCCESS, payload: {
-                token, user
-            }})
+            dispatch({
+                type: authConstants.LOGIN_SUCCESS, payload: {
+                    token, user
+                }
+            })
         } else {
-            if(res.status === 400) {
-                dispatch({type: authConstants.LOGIN_FAILURE, payload: {
-                    error: res.data.error
-                }})
+            if (res.status === 400) {
+                dispatch({
+                    type: authConstants.LOGIN_FAILURE, payload: {
+                        error: res.data.error
+                    }
+                })
             }
         }
 
@@ -29,24 +33,35 @@ export const login = (user) => {
 export const isUserLoggedIn = () => {
     return async dispatch => {
         const token = localStorage.getItem('token');
-        if(token) {
+        if (token) {
             const user = JSON.parse(localStorage.getItem('user'));
-            dispatch({type: authConstants.LOGIN_SUCCESS, payload: {
-                token, user
-            }})
+            dispatch({
+                type: authConstants.LOGIN_SUCCESS, payload: {
+                    token, user
+                }
+            })
         } else {
-            dispatch({type: authConstants.LOGIN_FAILURE, payload: {
-                error: 'Failed to login'
-            }})
+            dispatch({
+                type: authConstants.LOGIN_FAILURE, payload: {
+                    error: 'Failed to login'
+                }
+            })
         }
     }
 };
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+        dispatch({ type: authConstants.LOGOUT_REQUEST });
+        const res = await axios.post('/admin/signout');
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({ type: authConstants.LOGOUT_SUCCESS });
+        } else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE, payload: { error: res.data.error }
+            })
+        }
+
     }
 }
